Merge master 20180208

+++ tensorflow/python/keras/_impl/keras/estimator_test.py
@@ -23,43 +23,41 @@ import tempfile
 
 import numpy as np
 
+from tensorflow.python.estimator import run_config as run_config_lib
 from tensorflow.python.estimator.inputs import numpy_io
-from tensorflow.python.framework import dtypes
-from tensorflow.python.framework import ops
+from tensorflow.python.framework import test_util
 from tensorflow.python.keras._impl import keras
 from tensorflow.python.keras._impl.keras import testing_utils
+from tensorflow.python.keras._impl.keras.applications import mobilenet
 from tensorflow.python.platform import gfile
 from tensorflow.python.platform import test
+from tensorflow.python.summary.writer import writer_cache
+
 
 try:
   import h5py  # pylint:disable=g-import-not-at-top
 except ImportError:
   h5py = None
 
+_RANDOM_SEED = 1337
+_TRAIN_SIZE = 200
+_INPUT_SIZE = (10,)
+_NUM_CLASS = 2
+
 
 def simple_sequential_model():
   model = keras.models.Sequential()
-  model.add(
-      keras.layers.Conv2D(
-          32, kernel_size=(3, 3), activation='relu', input_shape=(14, 14, 3)))
-  model.add(keras.layers.MaxPooling2D(pool_size=(2, 2)))
-  model.add(keras.layers.Dropout(0.25))
-  model.add(keras.layers.Flatten())
-  model.add(keras.layers.Dense(16, activation='relu'))
-  model.add(keras.layers.Dropout(0.25))
-  model.add(keras.layers.Dense(3, activation='softmax'))
+  model.add(keras.layers.Dense(16, activation='relu', input_shape=_INPUT_SIZE))
+  model.add(keras.layers.Dropout(0.1))
+  model.add(keras.layers.Dense(_NUM_CLASS, activation='softmax'))
   return model
 
 
 def simple_functional_model():
-  a = keras.layers.Input(shape=(14, 14, 3))
-  b = keras.layers.Conv2D(32, kernel_size=(3, 3), activation='relu')(a)
-  b = keras.layers.MaxPooling2D(pool_size=(2, 2))(b)
-  b = keras.layers.Dropout(0.25)(b)
-  b = keras.layers.Flatten()(b)
-  b = keras.layers.Dense(16, activation='relu')(b)
-  b = keras.layers.Dropout(0.25)(b)
-  b = keras.layers.Dense(3, activation='softmax')(b)
+  a = keras.layers.Input(shape=_INPUT_SIZE)
+  b = keras.layers.Dense(16, activation='relu')(a)
+  b = keras.layers.Dropout(0.1)(b)
+  b = keras.layers.Dense(_NUM_CLASS, activation='softmax')(b)
   model = keras.models.Model(inputs=[a], outputs=[b])
   return model
 
@@ -70,33 +68,27 @@ def get_resource_for_simple_model(is_sequential, is_evaluate):
   if is_sequential:
     model.build()
   input_name = model.input_names[0]
-
-  np.random.seed(1337)
+  np.random.seed(_RANDOM_SEED)
   (x_train, y_train), (x_test, y_test) = testing_utils.get_test_data(
-      train_samples=200,
-      test_samples=100,
-      input_shape=(14, 14, 3),
-      num_classes=3)
+      train_samples=_TRAIN_SIZE,
+      test_samples=50,
+      input_shape=_INPUT_SIZE,
+      num_classes=_NUM_CLASS)
   y_train = keras.utils.to_categorical(y_train)
   y_test = keras.utils.to_categorical(y_test)
 
   train_input_fn = numpy_io.numpy_input_fn(
-      x={input_name: np.array(x_train, dtype=np.float32)},
-      y=np.array(y_train, dtype=np.float32),
+      x={input_name: x_train},
+      y=y_train,
       shuffle=False,
       num_epochs=None,
       batch_size=16)
 
   evaluate_input_fn = numpy_io.numpy_input_fn(
-      x={input_name: np.array(x_test, dtype=np.float32)},
-      y=np.array(y_test, dtype=np.float32),
-      num_epochs=1,
-      shuffle=False)
+      x={input_name: x_test}, y=y_test, num_epochs=1, shuffle=False)
 
   predict_input_fn = numpy_io.numpy_input_fn(
-      x={input_name: np.array(x_test, dtype=np.float32)},
-      num_epochs=1,
-      shuffle=False)
+      x={input_name: x_test}, num_epochs=1, shuffle=False)
 
   inference_input_fn = evaluate_input_fn if is_evaluate else predict_input_fn
 
@@ -106,9 +98,9 @@ def get_resource_for_simple_model(is_sequential, is_evaluate):
 
 def multi_inputs_multi_outputs_model():
   # test multi-input layer
-  a = keras.layers.Input(shape=(32,), name='input_a')
-  b = keras.layers.Input(shape=(32,), name='input_b')
-  dense = keras.layers.Dense(16, name='dense_1')
+  a = keras.layers.Input(shape=(16,), name='input_a')
+  b = keras.layers.Input(shape=(16,), name='input_b')
+  dense = keras.layers.Dense(8, name='dense_1')
   a_2 = dense(a)
   b_2 = dense(b)
   merged = keras.layers.concatenate([a_2, b_2], name='merge')
@@ -118,19 +110,26 @@ def multi_inputs_multi_outputs_model():
   model.compile(
       loss='categorical_crossentropy',
       optimizer='rmsprop',
-      metrics={'dense_2': 'accuracy',
-               'dense_3': 'accuracy'})
+      metrics={
+          'dense_2': 'categorical_accuracy',
+          'dense_3': 'categorical_accuracy'
+      })
   return model
 
 
-class TestKerasEstimator(test.TestCase):
+class TestKerasEstimator(test_util.TensorFlowTestCase):
 
   def setUp(self):
     self._base_dir = os.path.join(self.get_temp_dir(), 'keras_estimator_test')
     gfile.MakeDirs(self._base_dir)
+    self._config = run_config_lib.RunConfig(
+        tf_random_seed=_RANDOM_SEED, model_dir=self._base_dir)
 
   def tearDown(self):
-    gfile.DeleteRecursively(self._base_dir)
+    # Make sure nothing is stuck in limbo.
+    writer_cache.FileWriterCache.clear()
+    if os.path.isdir(self._base_dir):
+      gfile.DeleteRecursively(self._base_dir)
 
   def test_train(self):
     for is_sequential in [True, False]:
@@ -140,17 +139,19 @@ class TestKerasEstimator(test.TestCase):
       keras_model.compile(
           loss='categorical_crossentropy',
           optimizer='rmsprop',
-          metrics=['accuracy', 'mse', keras.metrics.categorical_accuracy])
+          metrics=['mse', keras.metrics.categorical_accuracy])
 
       with self.test_session():
         est_keras = keras.estimator.model_to_estimator(
-            keras_model=keras_model,
-            model_dir=tempfile.mkdtemp(dir=self._base_dir))
-        est_keras.train(input_fn=train_input_fn, steps=200 * 10 / 16)
-        eval_results = est_keras.evaluate(input_fn=eval_input_fn)
-        self.assertGreater(eval_results['accuracy'], 0.9)
-        self.assertGreater(eval_results['categorical_accuracy'], 0.9)
-        self.assertLess(eval_results['mse'], 0.1)
+            keras_model=keras_model, config=self._config)
+        before_eval_results = est_keras.evaluate(
+            input_fn=eval_input_fn, steps=1)
+        est_keras.train(input_fn=train_input_fn, steps=_TRAIN_SIZE / 16)
+        after_eval_results = est_keras.evaluate(input_fn=eval_input_fn, steps=1)
+        self.assertLess(after_eval_results['loss'], before_eval_results['loss'])
+
+      writer_cache.FileWriterCache.clear()
+      gfile.DeleteRecursively(self._config.model_dir)
 
   def test_evaluate(self):
     keras_model, (x_train, y_train), (
@@ -173,15 +174,17 @@ class TestKerasEstimator(test.TestCase):
 
     with self.test_session():
       keras_est = keras.estimator.model_to_estimator(
-          keras_model=keras_model,
-          model_dir=tempfile.mkdtemp(dir=self._base_dir))
+          keras_model=keras_model, config=self._config)
       est_eval = keras_est.evaluate(input_fn=eval_input_fn)
 
     metrics = ['loss'] + metrics
 
     # Check loss and all metrics match between keras and estimator.
     def shift(val):
-      return val / 10**int(log10(abs(val)))
+      if val == 0:
+        return 0
+      else:
+        return val / 10**int(log10(abs(val)))
 
     for i, metric_name in enumerate(metrics):
       self.assertAlmostEqual(
@@ -207,8 +210,7 @@ class TestKerasEstimator(test.TestCase):
 
     with self.test_session():
       keras_est = keras.estimator.model_to_estimator(
-          keras_model=keras_model,
-          model_dir=tempfile.mkdtemp(dir=self._base_dir))
+          keras_model=keras_model, config=self._config)
       est_pred = [
           np.argmax(y[keras_model.output_names[0]])
           for y in keras_est.predict(input_fn=pred_input_fn)
@@ -216,62 +218,41 @@ class TestKerasEstimator(test.TestCase):
     self.assertAllEqual(est_pred, keras_pred)
 
   def test_multi_inputs_multi_outputs(self):
-    np.random.seed(1337)
+    np.random.seed(_RANDOM_SEED)
     (a_train, c_train), (a_test, c_test) = testing_utils.get_test_data(
-        train_samples=200, test_samples=100, input_shape=(32,), num_classes=3)
+        train_samples=_TRAIN_SIZE,
+        test_samples=50,
+        input_shape=(16,),
+        num_classes=3)
+    np.random.seed(_RANDOM_SEED)
     (b_train, d_train), (b_test, d_test) = testing_utils.get_test_data(
-        train_samples=200, test_samples=100, input_shape=(32,), num_classes=2)
+        train_samples=_TRAIN_SIZE,
+        test_samples=50,
+        input_shape=(16,),
+        num_classes=2)
     c_train = keras.utils.to_categorical(c_train)
     c_test = keras.utils.to_categorical(c_test)
     d_train = keras.utils.to_categorical(d_train)
     d_test = keras.utils.to_categorical(d_test)
 
     def train_input_fn():
-      input_dict = {
-          'input_a':
-              ops.convert_to_tensor(
-                  np.array(a_train, dtype=np.float32), dtype=dtypes.float32),
-          'input_b':
-              ops.convert_to_tensor(
-                  np.array(b_train, dtype=np.float32), dtype=dtypes.float32)
-      }
-      output_dict = {
-          'dense_2':
-              ops.convert_to_tensor(
-                  np.array(c_train, dtype=np.float32), dtype=dtypes.float32),
-          'dense_3':
-              ops.convert_to_tensor(
-                  np.array(d_train, dtype=np.float32), dtype=dtypes.float32)
-      }
+      input_dict = {'input_a': a_train, 'input_b': b_train}
+      output_dict = {'dense_2': c_train, 'dense_3': d_train}
       return input_dict, output_dict
 
-    def evaluate_input_fn():
-      input_dict = {
-          'input_a':
-              ops.convert_to_tensor(
-                  np.array(a_test, dtype=np.float32), dtype=dtypes.float32),
-          'input_b':
-              ops.convert_to_tensor(
-                  np.array(b_test, dtype=np.float32), dtype=dtypes.float32)
-      }
-      output_dict = {
-          'dense_2':
-              ops.convert_to_tensor(
-                  np.array(c_test, dtype=np.float32), dtype=dtypes.float32),
-          'dense_3':
-              ops.convert_to_tensor(
-                  np.array(d_test, dtype=np.float32), dtype=dtypes.float32)
-      }
+    def eval_input_fn():
+      input_dict = {'input_a': a_test, 'input_b': b_test}
+      output_dict = {'dense_2': c_test, 'dense_3': d_test}
       return input_dict, output_dict
 
     with self.test_session():
       model = multi_inputs_multi_outputs_model()
       est_keras = keras.estimator.model_to_estimator(
-          keras_model=model, model_dir=tempfile.mkdtemp(dir=self._base_dir))
-      est_keras.train(input_fn=train_input_fn, steps=200 * 10 / 16)
-      eval_results = est_keras.evaluate(input_fn=evaluate_input_fn, steps=1)
-      self.assertGreater(eval_results['accuracy_dense_2'], 0.5)
-      self.assertGreater(eval_results['accuracy_dense_3'], 0.5)
+          keras_model=model, config=self._config)
+      before_eval_results = est_keras.evaluate(input_fn=eval_input_fn, steps=1)
+      est_keras.train(input_fn=train_input_fn, steps=_TRAIN_SIZE / 16)
+      after_eval_results = est_keras.evaluate(input_fn=eval_input_fn, steps=1)
+      self.assertLess(after_eval_results['loss'], before_eval_results['loss'])
 
   def test_init_from_file(self):
     if h5py is None:
@@ -285,7 +266,7 @@ class TestKerasEstimator(test.TestCase):
       keras_model.compile(
           loss='categorical_crossentropy',
           optimizer='rmsprop',
-          metrics=['accuracy'])
+          metrics=['categorical_accuracy'])
       keras_model.fit(x_train, y_train, epochs=1)
       keras_pred = [np.argmax(y) for y in keras_model.predict(x_test)]
       fname = os.path.join(self._base_dir, 'keras_model.h5')
@@ -293,8 +274,7 @@ class TestKerasEstimator(test.TestCase):
 
     with self.test_session():
       keras_est = keras.estimator.model_to_estimator(
-          keras_model_path=fname,
-          model_dir=tempfile.mkdtemp(dir=self._base_dir))
+          keras_model_path=fname, config=self._config)
       est_pred = [
           np.argmax(y[keras_model.output_names[0]])
           for y in keras_est.predict(input_fn=pred_input_fn)
@@ -324,39 +304,27 @@ class TestKerasEstimator(test.TestCase):
             keras_model_path='gs://bucket/object')
 
   def test_invalid_ionames_error(self):
-    np.random.seed(1337)
     (x_train, y_train), (_, _) = testing_utils.get_test_data(
-        train_samples=200, test_samples=100, input_shape=(10,), num_classes=2)
+        train_samples=_TRAIN_SIZE,
+        test_samples=100,
+        input_shape=(10,),
+        num_classes=2)
     y_train = keras.utils.to_categorical(y_train)
 
     def invald_input_name_input_fn():
-      input_dict = {
-          'invalid_input_name':
-              ops.convert_to_tensor(
-                  np.array(x_train, dtype=np.float32), dtype=dtypes.float32),
-      }
-      output = ops.convert_to_tensor(
-          np.array(y_train, dtype=np.float32), dtype=dtypes.float32)
-      return input_dict, output
+      input_dict = {'invalid_input_name': x_train}
+      return input_dict, y_train
 
     def invald_output_name_input_fn():
-      input_dict = {
-          'input_1':
-              ops.convert_to_tensor(
-                  np.array(x_train, dtype=np.float32), dtype=dtypes.float32),
-      }
-      output_dict = {
-          'invalid_output_name':
-              ops.convert_to_tensor(
-                  np.array(y_train, dtype=np.float32), dtype=dtypes.float32),
-      }
+      input_dict = {'input_1': x_train}
+      output_dict = {'invalid_output_name': y_train}
       return input_dict, output_dict
 
     model = simple_functional_model()
     model.compile(
         loss='categorical_crossentropy', optimizer='adam', metrics=['acc'])
     est_keras = keras.estimator.model_to_estimator(
-        keras_model=model, model_dir=tempfile.mkdtemp(dir=self._base_dir))
+        keras_model=model, config=self._config)
 
     with self.test_session():
       with self.assertRaises(ValueError):
@@ -366,26 +334,23 @@ class TestKerasEstimator(test.TestCase):
         est_keras.train(input_fn=invald_output_name_input_fn, steps=100)
 
   def test_custom_objects(self):
-    keras_model, (_, _), (
-        _, _), train_input_fn, eval_input_fn = get_resource_for_simple_model(
-            is_sequential=True, is_evaluate=True)
-
-    class CustomOp(keras.optimizers.RMSprop):
-      pass
-
-    def custom_loss(y_true, y_pred):
-      return keras.losses.categorical_crossentropy(y_true, y_pred)
-
-    keras_model.compile(
-        loss=custom_loss, optimizer=CustomOp(), metrics=['accuracy'])
+    keras_mobile = mobilenet.MobileNet(weights=None)
+    keras_mobile.compile(loss='categorical_crossentropy', optimizer='adam')
+    custom_objects = {
+        'relu6': mobilenet.relu6,
+        'DepthwiseConv2D': mobilenet.DepthwiseConv2D
+    }
+    with self.assertRaisesRegexp(ValueError, 'relu6'):
+      with self.test_session():
+        keras.estimator.model_to_estimator(
+            keras_model=keras_mobile,
+            model_dir=tempfile.mkdtemp(dir=self._base_dir))
 
     with self.test_session():
-      est_keras = keras.estimator.model_to_estimator(
-          keras_model=keras_model,
-          model_dir=tempfile.mkdtemp(dir=self._base_dir))
-      est_keras.train(input_fn=train_input_fn, steps=200 * 10 / 16)
-      eval_results = est_keras.evaluate(input_fn=eval_input_fn)
-      self.assertGreater(eval_results['accuracy'], 0.9)
+      keras.estimator.model_to_estimator(
+          keras_model=keras_mobile,
+          model_dir=tempfile.mkdtemp(dir=self._base_dir),
+          custom_objects=custom_objects)
 
 
 if __name__ == '__main__':
